+++ test/unit/index.js
@@ -17,6 +17,7 @@ var pathFiles = [
 	'./helpers/slots.js',
 	'./helpers/wsApi.js',
 	'./helpers/z_schema.js',
+	'./helpers/sort_by.js',
 
 	'./logic/account.js',
 	'./logic/blockReward.js', // @slow
